# Blogging Platform/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Laravel Medium Clone API",
        "description": "API documentation for the Laravel Medium Clone project",
        "contact": {
            "email": "admin@example.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "Local API Server"
        }
    ],
    "paths": {
...
        "/clap/{post}": {
            "post": {
                "tags": [
                    "Posts"
                ],
                "summary": "Clap or unclap a post",
                "operationId": "8a1d6317e07aa26effa6c66e48c6268e",
                "parameters": [
                    {
                        "name": "post",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successfully toggled clap status",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "clapsCount": {
                                            "type": "integer",
                                            "example": 10
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "csrf_token": []
                    }
                ]
            }
        },
        "/follow/{user}": {
            "post": {
                "tags": [
                    "Followers"
                ],
                "summary": "Follow or unfollow a user",
                "operationId": "e6994afd4300c37a8070183e07e8fa2f",
                "parameters": [
                    {
                        "name": "user",
                        "in": "path",
                        "description": "The ID of the user to follow/unfollow",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successfully toggled follow status",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "followersCount": {
                                            "type": "integer",
                                            "example": 5
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "csrf_token": []
                    }
                ]
            }
        },
...
        "/post/create": {
            "post": {
                "tags": [
                    "Posts"
                ],
                "summary": "Store a newly created post",
                "operationId": "99ac18525eda9ec469e45236535ae5c1",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "title",
                                    "body",
                                    "category_id"
                                ],
                                "properties": {
                                    "title": {
                                        "type": "string",
                                        "example": "My First Post"
                                    },
                                    "body": {
                                        "type": "string",
                                        "example": "Content of the post..."
                                    },
                                    "category_id": {
                                        "type": "integer",
                                        "example": 1
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Post created successfully"
                    },
                    "422": {
                        "description": "Validation error"
                    }
                },
                "security": [
                    {
                        "csrf_token": []
                    }
                ]
            }
        },
...
        "/post/{post}": {
            "put": {
                "tags": [
                    "Posts"
                ],
                "summary": "Update the specified post",
                "operationId": "ef3d25726bac3f85b425951d0f38b6c2",
                "parameters": [
                    {
                        "name": "post",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "title": {
                                        "type": "string"
                                    },
                                    "body": {
                                        "type": "string"
                                    },
                                    "category_id": {
                                        "type": "integer"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Post updated successfully"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "Post not found"
                    }
                },
                "security": [
                    {
                        "csrf_token": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Posts"
                ],
                "summary": "Remove the specified post",
                "operationId": "e1a1b68b72d67133153743fb62cb6a61",
                "parameters": [
                    {
                        "name": "post",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Post deleted successfully"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "Post not found"
                    }
                },
                "security": [
                    {
                        "csrf_token": []
                    }
                ]
            }
        },
        "/category/{category}": {
            "get": {
                "tags": [
                    "Posts"
                ],
                "summary": "Display posts by category",
                "operationId": "dfb1e7bc32928c7b43b4d3c63be5abe8",
                "parameters": [
                    {
                        "name": "category",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    }
                }
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "csrf_token": {
                "type": "apiKey",
                "description": "1. Open http://localhost:8000/csrf-token in a browser. 2. Copy the value of 'token'. 3. Click 'Authorize' here and paste that value.",
                "name": "X-CSRF-TOKEN",
                "in": "header"
            }
        }
    },
    "security": [
        {
            "csrf_token": []
        }
    ],
...
}